fix(mcp): extend log redaction to McpOAuthCredentialStore
PR #1970 changes only in-memory identity fallback logic and adds no
logging, so it does not introduce a leak on its own. But an audit for
the same issue found McpOAuthCredentialStore has the same shape of
gap: it also logs a raw exception in its credentials-file load path.
The secrets file is decrypted before that callback runs, so a
malformed-JSON or decryption exception could in principle echo a
fragment of plaintext credential content into the log.

Move RedactForLogging out of McpClientManager and into
SecretOutputRedactor, so both McpClientManager and
McpOAuthCredentialStore route their exception logging through the
same shared helper.
